+++ src/main/java/net/succ/solar_punk/ModClientEvents.java
@@ -30,6 +30,7 @@ import net.succ.solar_punk.client.model.SolarPowerTowerModel;
 import net.succ.solar_punk.client.model.TurbineCasingGlassModel;
 import net.succ.solar_punk.client.model.TurbineCasingModel;
 import net.succ.solar_punk.client.renderer.AndesiteSolarPanelRenderer;
+import net.succ.solar_punk.client.renderer.BiomassGasifierRenderer;
 import net.succ.solar_punk.client.renderer.BiofilterRenderer;
 import net.succ.solar_punk.client.renderer.FermentationVatRenderer;
 import net.succ.solar_punk.client.renderer.GeyserCapRenderer;
@@ -96,6 +97,7 @@ public class ModClientEvents {
 
     @SubscribeEvent
     public static void onRegisterRenderers(EntityRenderersEvent.RegisterRenderers event) {
+        event.registerBlockEntityRenderer(ModBlockEntities.BIOMASS_GASIFIER.get(), BiomassGasifierRenderer::new);
         event.registerBlockEntityRenderer(ModBlockEntities.GEYSER_CAP.get(), GeyserCapRenderer::new);
         event.registerBlockEntityRenderer(ModBlockEntities.TURBINE_ROTOR.get(), TurbineRotorRenderer::new);
         event.registerBlockEntityRenderer(ModBlockEntities.ANDESITE_SOLAR_PANEL.get(), AndesiteSolarPanelRenderer::new);
@@ -112,6 +114,10 @@ public class ModClientEvents {
             ModSpriteShifts.init();
             PonderIndex.addPlugin(new SolarPunkPonderPlugin());
             registerTooltips();
+
+            SimpleBlockEntityVisualizer.builder(ModBlockEntities.BIOMASS_GASIFIER.get())
+                    .factory(SingleAxisRotatingVisual::shaft)
+                    .apply();
 
             SimpleBlockEntityVisualizer.builder(ModBlockEntities.TURBINE_ROTOR.get())
                     .factory(SingleAxisRotatingVisual::shaft)